fastapi-backend/requirements.txt
# Core Web & API
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# OpenAI API (Whisper + Embeddings)
openai>=1.51.0

# Utilities
numpy<2.0.0
python-dotenv>=1.0.1

# Misc
pydantic>=2.9.0
python-multipart>=0.0.9
pymongo>=4.8.0